+++ app/index.js
@@ -9,7 +9,7 @@ const container = ConfigureContainer();
 
 (async () => {
     try {
-        const argv = Yargs(hideBin(process.argv))
+        const { argv } = Yargs(hideBin(process.argv))
             .usage(
                 "Usage: $0 -GITLAB_API_ENDPOINT [string] " +
                     "-GITLAB_PERSONAL_TOKEN [string] -GITLAB_PROJECT_ID [string] " +
@@ -18,14 +18,7 @@ const container = ConfigureContainer();
                     "-ISSUE_CLOSED_SECONDS [num] " +
                     "-TZ [string] -NODE_ENV [string]"
             )
-            .demandOption(["GITLAB_PERSONAL_TOKEN", "GITLAB_PROJECT_ID"]).argv;
-
-        const defaultOptions = {
-            GITLAB_API_ENDPOINT: "https://gitlab.com/api/v4",
-            NODE_ENV: process.env.NODE_ENV,
-            TZ: "Australia/Melbourne",
-            SERVICE_PROVIDER: Constants.SERVICE_PROVIDER_GITLAB
-        };
+            .demandOption(["GITLAB_PERSONAL_TOKEN", "GITLAB_PROJECT_ID"]);
 
         const {
             GITLAB_API_ENDPOINT,
@@ -40,22 +33,18 @@ const container = ConfigureContainer();
             TZ,
             NODE_ENV
         } = argv;
-        if (!GITLAB_PROJECT_ID) throw Error("GitLab project id is required");
-        if (!GITLAB_PERSONAL_TOKEN) throw Error("Gitlab personal token is required");
-        const config = {
-            GITLAB_API_ENDPOINT,
-            GITLAB_PERSONAL_TOKEN,
-            GITLAB_PROJECT_ID,
-
-            TARGET_BRANCH,
-            TARGET_TAG_REGEX,
-            ISSUE_CLOSED_SECONDS,
-            SERVICE_PROVIDER,
-
-            TZ,
-            NODE_ENV,
-            ...defaultOptions
-        };
+        if (!GITLAB_PROJECT_ID) throw new Error("GitLab project id is required");
+        if (!GITLAB_PERSONAL_TOKEN) throw new Error("Gitlab personal token is required");
+        const config = { ...Constants.defaultOptions };
+        if (GITLAB_API_ENDPOINT) config.GITLAB_API_ENDPOINT = GITLAB_API_ENDPOINT;
+        if (GITLAB_PERSONAL_TOKEN) config.GITLAB_PERSONAL_TOKEN = GITLAB_PERSONAL_TOKEN;
+        if (GITLAB_PROJECT_ID) config.GITLAB_PROJECT_ID = GITLAB_PROJECT_ID;
+        if (TARGET_BRANCH) config.TARGET_BRANCH = TARGET_BRANCH;
+        if (TARGET_TAG_REGEX) config.TARGET_TAG_REGEX = TARGET_TAG_REGEX;
+        if (ISSUE_CLOSED_SECONDS) config.ISSUE_CLOSED_SECONDS = ISSUE_CLOSED_SECONDS;
+        if (SERVICE_PROVIDER) config.SERVICE_PROVIDER = SERVICE_PROVIDER;
+        if (TZ) config.TZ = TZ;
+        if (NODE_ENV) config.NODE_ENV = NODE_ENV;
         container.register({ config: asValue(config) });
         const GitLabReleaseNoteGenerator = container.resolve("gitLabReleaseNoteGenerator");
         await GitLabReleaseNoteGenerator.run();
