+++ app/publishers/index.js
