+++ app/services/logger.js
@@ -4,23 +4,24 @@ const Moment = require("moment-timezone");
 module.exports = class LoggerService {
     constructor({ config }) {
         this.config = config;
+        this.console = console;
         if (this.config.NODE_ENV === "test") this.console.log = () => {};
     }
 
     debug(message) {
         // eslint-disable-next-line no-console
-        console.log(Chalk.whiteBright(`${Moment.tz(this.config.TZ).format()} [DEBUG] ${message}`));
+        this.console.log(Chalk.whiteBright(`${Moment.tz(this.config.TZ).format()} [DEBUG] ${message}`));
     }
     info(message) {
         // eslint-disable-next-line no-console
-        console.log(Chalk.greenBright(`${Moment.tz(this.config.TZ).format()} [INFO] ${message}`));
+        this.console.log(Chalk.greenBright(`${Moment.tz(this.config.TZ).format()} [INFO] ${message}`));
     }
     warn(message) {
         // eslint-disable-next-line no-console
-        console.log(Chalk.yellowBright(`${Moment.tz(this.config.TZ).format()} [WARN] ${message}`));
+        this.console.log(Chalk.yellowBright(`${Moment.tz(this.config.TZ).format()} [WARN] ${message}`));
     }
     error(message) {
         // eslint-disable-next-line no-console
-        console.log(Chalk.redBright(`${Moment.tz(this.config.TZ).format()} [ERROR] ${message}`));
+        this.console.log(Chalk.redBright(`${Moment.tz(this.config.TZ).format()} [ERROR] ${message}`));
     }
 };
